adding few commands
