--- cmd/useAccount.go
@@ -1,89 +1,0 @@
-/*
-Copyright © 2021 NAME HERE <EMAIL ADDRESS>
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package cmd
-
-import (
-	"fmt"
-	config "github.com/litmuschaos/litmusctl/pkg/config"
-	"github.com/litmuschaos/litmusctl/pkg/types"
-	"github.com/spf13/cobra"
-	"os"
-)
-
-// useAccountCmd represents the useAccount command
-var useAccountCmd = &cobra.Command{
-	Use:   "use-account",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		username, err  := cmd.Flags().GetString("username")
-		if err != nil{
-			fmt.Print(err)
-			os.Exit(1)
-		}
-
-		endpoint, err := cmd.Flags().GetString("endpoint")
-		if err != nil{
-			fmt.Print(err)
-			os.Exit(1)
-		}
-		defaultFileName := types.DefaultFileName
-		exists := config.FileExists(defaultFileName)
-
-		err = config.ConfigSyntaxCheck(defaultFileName)
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
-		}
-
-
-		if exists {
-			err = config.UpdateCurrent(types.Current{
-				CurrentAccount: endpoint,
-				CurrentUser: username,
-			}, defaultFileName)
-			if err != nil {
-				fmt.Println(err)
-				os.Exit(1)
-			}
-
-		} else {
-			fmt.Println("File Not exists")
-		}
-
-	},
-}
-
-func init() {
-	configCmd.AddCommand(useAccountCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	//useAccountCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	useAccountCmd.Flags().StringP("username", "u", "", "Help message for toggle")
-	useAccountCmd.Flags().StringP("endpoint", "e", "", "Help message for toggle")
-
-}
