+++ pkg/types/api_types.go
@@ -1,6 +1,5 @@
 package types
 
-
 type AuthResponse struct {
 	AccessToken string `json:"access_token"`
 	ExpiresIn   int64    `json:"expires_in"`
