+++ pkg/types/config_types.go
@@ -25,5 +25,11 @@ type Current struct {
 	CurrentUser string `yaml:"current-user",json:"current-user"`
 }
 
+type UpdateLitmusCtlConfig struct {
+	CurrentAccount string `yaml:"current-account",json:"current-account"`
+	CurrentUser string `yaml:"current-user",json:"current-user"`
+	Account 	 Account `yaml:"account",json:"account"`
+}
+
 const DefaultFileName = "litmusconfig.yaml"
